moving content from connect to column
